+++ Modules/QuestionBank/routes/web.php
@@ -19,17 +19,21 @@ use Modules\QuestionBank\Http\Controllers\WeakTopicSessionController;
 | Add server-rendered pages here; API lives in routes/api.php.
 */
 
-Route::middleware(['auth', 'learner'])->group(function (): void {
+Route::middleware(['auth', 'learner', 'permission:question.view'])->group(function (): void {
     Route::get('/qbank', QuestionBankPageController::class)->name('qbank.index');
-    Route::get('/qbank/bookmarks', [QuestionBookmarkPageController::class, 'index'])->name('qbank.bookmarks');
+    Route::get('/qbank/bookmarks', [QuestionBookmarkPageController::class, 'index'])
+        ->middleware('permission:bookmark.view')->name('qbank.bookmarks');
     Route::delete('/qbank/bookmarks/{question}', [QuestionBookmarkPageController::class, 'destroy'])
+        ->middleware('permission:bookmark.delete')
         ->name('qbank.bookmarks.destroy');
     Route::post('/qbank/bookmarks/session', [QuestionBookmarkPageController::class, 'startSession'])
+        ->middleware('permission:session.create')
         ->name('qbank.bookmarks.session');
-    Route::get('/qbank/create', [CustomSessionController::class, 'create'])->name('qbank.create');
-    Route::post('/qbank/create', [CustomSessionController::class, 'store'])->name('qbank.store');
-    Route::post('/qbank/create/count', [CustomSessionController::class, 'count'])->name('qbank.count');
+    Route::get('/qbank/create', [CustomSessionController::class, 'create'])->middleware('permission:session.create')->name('qbank.create');
+    Route::post('/qbank/create', [CustomSessionController::class, 'store'])->middleware('permission:session.create')->name('qbank.store');
+    Route::post('/qbank/create/count', [CustomSessionController::class, 'count'])->middleware('permission:session.create')->name('qbank.count');
     Route::post('/qbank/weak-topics/{lesson}/session', WeakTopicSessionController::class)
+        ->middleware('permission:session.create')
         ->name('qbank.weak-topics.session');
 
     Route::prefix('qbank/taxonomy/lookups')->name('qbank.taxonomy.lookups.')->group(function (): void {
@@ -43,16 +47,16 @@ Route::middleware(['auth', 'learner'])->group(function (): void {
         Route::get('/tags', [TaxonomyLookupController::class, 'tags'])->name('tags');
     });
 
-    Route::get('/qbank/session/{session}', [StudySessionController::class, 'show'])->name('qbank.session');
-    Route::post('/qbank/session/{session}/answer', [StudySessionController::class, 'answer'])->name('qbank.session.answer');
-    Route::post('/qbank/session/{session}/annotation', [StudySessionController::class, 'annotate'])->name('qbank.session.annotate');
-    Route::post('/qbank/session/{session}/feedback', [QuestionFeedbackController::class, 'store'])->name('qbank.session.feedback');
-    Route::post('/qbank/session/{session}/pause', [StudySessionController::class, 'pause'])->name('qbank.session.pause');
-    Route::post('/qbank/session/{session}/resume', [StudySessionController::class, 'resume'])->name('qbank.session.resume');
-    Route::post('/qbank/session/{session}/finish', [StudySessionController::class, 'finish'])->name('qbank.session.finish');
-    Route::patch('/qbank/session/{session}/name', [SessionHistoryController::class, 'rename'])->name('qbank.session.rename');
-    Route::post('/qbank/session/{session}/repeat', [SessionHistoryController::class, 'repeat'])->name('qbank.session.repeat');
-    Route::delete('/qbank/session/{session}', [SessionHistoryController::class, 'destroy'])->name('qbank.session.destroy');
-    Route::get('/qbank/session/{session}/summary', SessionSummaryController::class)->name('qbank.summary');
-    Route::get('/qbank/session/{session}/review', QuestionReviewController::class)->name('qbank.review');
+    Route::get('/qbank/session/{session}', [StudySessionController::class, 'show'])->middleware('permission:session.start')->name('qbank.session');
+    Route::post('/qbank/session/{session}/answer', [StudySessionController::class, 'answer'])->middleware('permission:session.start')->name('qbank.session.answer');
+    Route::post('/qbank/session/{session}/annotation', [StudySessionController::class, 'annotate'])->middleware('permission:session.submit')->name('qbank.session.annotate');
+    Route::post('/qbank/session/{session}/feedback', [QuestionFeedbackController::class, 'store'])->middleware('permission:session.submit')->name('qbank.session.feedback');
+    Route::post('/qbank/session/{session}/pause', [StudySessionController::class, 'pause'])->middleware('permission:session.submit')->name('qbank.session.pause');
+    Route::post('/qbank/session/{session}/resume', [StudySessionController::class, 'resume'])->middleware('permission:session.submit')->name('qbank.session.resume');
+    Route::post('/qbank/session/{session}/finish', [StudySessionController::class, 'finish'])->middleware('permission:session.submit')->name('qbank.session.finish');
+    Route::patch('/qbank/session/{session}/name', [SessionHistoryController::class, 'rename'])->middleware('permission:session.submit')->name('qbank.session.rename');
+    Route::post('/qbank/session/{session}/repeat', [SessionHistoryController::class, 'repeat'])->middleware('permission:session.repeat')->name('qbank.session.repeat');
+    Route::delete('/qbank/session/{session}', [SessionHistoryController::class, 'destroy'])->middleware('permission:session.delete')->name('qbank.session.destroy');
+    Route::get('/qbank/session/{session}/summary', SessionSummaryController::class)->middleware('permission:session.review')->name('qbank.summary');
+    Route::get('/qbank/session/{session}/review', QuestionReviewController::class)->middleware('permission:session.review')->name('qbank.review');
 });
